+++ src/Idasen.SystemTray.Win11/Interfaces/ISitStandReminderManager.cs
@@ -1,0 +1,7 @@
+namespace Idasen.SystemTray.Win11.Interfaces ;
+
+public interface ISitStandReminderManager : IDisposable
+{
+    void Initialize ( ) ;
+    void HandleNotificationAction ( string action ) ;
+}